pyproject.toml
[tool.poetry]
name = "biochatter-server"
version = "0.5.4"
description = "FastAPI server to provide RESTful API for biochatter"
authors = [
    "fengsh <shaohong.feng.78@gmail.com>",
    "Sebastian Lobentanzer <sebastian.lobentanzer@gmail.com>"
]
license = "MIT"
packages = [
    { include = "src" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
Flask = "^3.0.0"
python-dotenv = "^1.0.0"
schedule = "^1.2.1"
gunicorn = "^21.2.0"
neo4j-utils = "^0.0.7"
fastapi = "^0.111.0"
langchain = "0.2.5"
langchain-community = "0.2.5"
langchain-openai = "0.1.14"
biochatter = "0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
bump2version = "^1.0.1"
coverage = "^7.6.0"
coverage-badge = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "unittests",
]